Update sharebar to latest devpack, remove Omniture
This looks like a big change, but it was mostly running `npm run provision`,
followed by updating the out of date modules, followed by cleaning up any
new lint errors.

There was also some work to strip out the component-omniture which should
be deprecated, as we are moving toward generic instrumentation.

This is BREAKING CHANGE, due to the above.